Insercción de menus y algunas funcionalidades que hacen lo pedido correctamente insertadas

+++ tarea3dwes/src/main/java/com/pelayora/tarea3dwes/repositorios/EjemplarRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.pelayora.tarea3dwes.modelo.Ejemplar;
 
 @Repository
-public interface EjemplarRepository extends JpaRepository<Ejemplar, Integer> {
+public interface EjemplarRepository extends JpaRepository<Ejemplar, Long> {
 
 }
